// packages/server/src/lib/database.d.ts
declare const prisma: PrismaClient<
	{
		log: ("error" | "warn" | "query")[];
	},
	"error" | "warn" | "query",
	import("@prisma/client/runtime/client").DefaultArgs
>;
// ...
export declare const tokenService: {
	create(data: { token: string; userId: string; expiresAt: Date }): Promise<{
		token: string;
		id: string;
		userId: string;
		createdAt: Date;
		expiresAt: Date;
		revoked: boolean;
	}>;
	findByToken(token: string): Promise<
		| ({
				user: {
					id: string;
					username: string;
					role: string;
					createdAt: Date;
					passwordHash: string;
					updatedAt: Date;
				};
		  } & {
				token: string;
				id: string;
				userId: string;
				createdAt: Date;
				expiresAt: Date;
				revoked: boolean;
		  })
		| null
	>;
	revoke(token: string): Promise<{
		token: string;
		id: string;
		userId: string;
		createdAt: Date;
		expiresAt: Date;
		revoked: boolean;
	}>;
	revokeAllByUser(userId: string): Promise<import("@prisma/client").Prisma.BatchPayload>;
	deleteExpired(): Promise<import("@prisma/client").Prisma.BatchPayload>;
};